Added navbar,signup and login page used material ui for inputs and buttons Functions are same as v1 Made UI same as that of landing page Responsive
