+++ Client-React/v1/public/manifest.json
