+++ Client-React/v2/public/manifest.json
