+++ Client-React/v2/src/page/login.js
@@ -1,37 +1,48 @@
 import React, { useState } from "react";
 import { Bounce, Zoom } from "react-reveal";
-import TextField from "@mui/material/TextField";
+import TextField
+   from "@mui/material/TextField";
+import Button from "@mui/material/Button";
+
+import { useEffect } from "react";
+import axios from "axios";
+import Navbar from "./dashboard";
 const Login = () => {
+ 
   const [email, setemail] = useState("");
   const [password, setpassword] = useState("");
+ 
 
-  const SubmitHandler = (e) => {
+  const SubmitHandler = async (e) => {
     e.preventDefault();
     //First check if password and cpassword are same
-    if (!email || !password) {
+    if ( !email || !password ) {
       alert("Please fill all the fields");
       return;
     }
-
+    
     alert("Login done");
-    //Call API to signup
 
+    
     setemail("");
     setpassword("");
+   
   };
   return (
     <>
-      <div className="container">
+    <Navbar l2={true}/>
+      <div className="container signup1">
         <div className="signup row">
           {/*Entry part */}
-          <h6 className="text-center header ">Login</h6>
+          <h1 className="text-center header ">Login</h1>
           <br />
           <div className="col-md-6 col-lg-6 col-12 d-flex justify-content-center ">
             <Bounce>
               <form action="POST" className="inputsection">
-                <div className="inputwrap"></div>
+               
                 <div className="inputwrap">
                   <TextField
+                    
                     id="outlined-basic"
                     label="Enter your email"
                     variant="outlined"
@@ -45,6 +56,7 @@ const Login = () => {
                 </div>
                 <div className="inputwrap">
                   <TextField
+                    
                     id="outlined-basic"
                     label="Enter your password"
                     variant="outlined"
@@ -55,15 +67,24 @@ const Login = () => {
                     required
                   />
                 </div>
-
-                <div className="btnwrap" onClick={SubmitHandler}>
-                  <button className="text-center btn">Sign-up</button>
+               
+                <div className="btnwrap">
+                  <Button
+                    style={{ width: "100%" }}
+                    onClick={SubmitHandler}
+                    variant="contained"
+                    size="large"
+                    className=""
+                  >
+                    Login
+                  </Button>
+                 
                 </div>
               </form>
             </Bounce>
           </div>
           {/*Image Part */}
-          <div className="col-md-6 col-lg-6 col-12 ">
+          <div className="col-md-6 col-lg-6 col-12 d-flex justify-content-center flexer mm">
             <Zoom>
               <img
                 src="https://res.cloudinary.com/vigneshshettyin/image/upload/v1631588908/oia0inntihtas3ymsvgi.png"
@@ -71,9 +92,7 @@ const Login = () => {
                 className="img-fluid"
               />
             </Zoom>
-            <h6 className="text-center">
-              <strong>Just login step and then we can taste your URLs</strong>
-            </h6>
+            
           </div>
         </div>
       </div>
