+++ Client-React/v2/src/styles/loading-style.js
