+++ Client-React/v2/src/styles/signup-login-styles.css
@@ -1,0 +1,88 @@
+/*Main signup container*/
+.signup{
+    height: 80vh;
+    border: 5px solid black;
+    border-radius: 0.5rem;
+    margin: 0.1rem;
+    margin-top: 0.8rem !important;
+    margin-bottom: 0.8rem !important;
+    background-color: #fff;
+  
+
+}
+.signup1{
+    border: 5px solid rgb(0, 0, 0);
+    border-radius: 0.5rem;
+   
+}
+.inputwrap{
+    margin: 0.5rem;
+    padding: 0.5rem;
+    border: 2px solid black;
+    border-radius: 0.5rem;
+}
+.btnwrap{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+.btnwrap Button{
+    background-color: rgb(0, 0, 0) !important;
+    width: 12rem !important;
+    border-radius: 0.5rem;
+}
+.btnwrap Button:hover{
+    background-color: #fff !important ;
+    color: black;
+    border: 2px solid black;
+}
+
+.flexer{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    margin-bottom: 5rem;
+    
+}
+.mm{
+    margin-bottom: 8rem;
+}
+.navwrap{
+padding: 0.5rem;
+margin: 0.3rem;
+border: 2px solid black;
+border-radius: 0.5rem;
+color: #fff;
+}
+
+.navwrap .navbtn{
+    border: 2px solid black;
+    margin: 0.3rem;
+    border-radius: 0.4rem;
+    color: #fff;
+    background-color: black;
+    transition: all 1s;
+    cursor: pointer;
+    
+}
+ .navbtn:hover{
+    background-color: #fff !important;
+    color: black !important;
+}
+
+/**Styling for link*/
+.link{
+    text-decoration: none;
+    color: #fff;
+}
+
+.link:hover{
+    color: black;
+}
+
+/*To tell whcih site is active*/
+.active{
+text-decoration: underline;
+}
+
